+++ src/lib/Config.ts
@@ -28,6 +28,7 @@ export interface Config {
   ViewOnly: boolean;
   HideBalances: boolean;
   BudgetSplitMin: number;
+  BuySignalThreshold: number;
   /**
    * When enabled - the system simulates bying and selling, without actually doing that on the exchange.
    * It also does not check if the stable coin balance is real or not.